+++ azplugins/PairPotentials.cu
@@ -48,5 +48,10 @@ template cudaError_t compute_pair_potential<azplugins::detail::PairEvaluatorShif
     (const pair_args_t& pair_args,
      const typename azplugins::detail::PairEvaluatorShiftedLJ::param_type *d_params);
 
+//! Kernel driver for Spline pair potential
+template cudaError_t compute_pair_potential<azplugins::detail::PairEvaluatorSpline>
+    (const pair_args_t& pair_args,
+     const typename azplugins::detail::PairEvaluatorSpline::param_type *d_params);
+
 } // end namespace gpu
 } // end namespace azplugins